requirements.txt
python-decouple
requests
websockets